+++ src/state.ts
@@ -11,7 +11,8 @@ import {
   Form,
   FormDefinition,
   ValidationResponse,
-  GroupDefinition
+  GroupDefinition,
+  ErrorFunc
 } from "./form";
 import {
   deepCopy,
@@ -26,6 +27,10 @@ import { RepeatingFormAccessor } from "./repeating-form-accessor";
 import { RepeatingFormIndexedAccessor } from "./repeating-form-indexed-accessor";
 import { FormAccessorBase } from "./form-accessor-base";
 import { ValidateOptions } from "./validate-options";
+import {
+  StateConverterOptions,
+  StateConverterOptionsWithContext
+} from "./converter";
 
 export interface FieldAccessorAllows {
   (fieldAccessor: FieldAccessor<any, any>): boolean;
@@ -83,6 +88,8 @@ export interface FormStateOptions<M> {
   update?: UpdateFunc<any, any>;
 
   context?: any;
+  converterOptions?: StateConverterOptions;
+  requiredError?: string | ErrorFunc;
 }
 
 export type SaveStatusOptions = "before" | "rightAfter" | "after";
@@ -117,6 +124,8 @@ export class FormState<
   updateFunc: UpdateFunc<any, any> | null;
 
   _context: any;
+  _converterOptions: StateConverterOptions;
+  _requiredError: string | ErrorFunc;
 
   constructor(
     public form: Form<M, D, G>,
@@ -166,6 +175,8 @@ export class FormState<
       this.blurFunc = null;
       this.updateFunc = null;
       this._context = undefined;
+      this._converterOptions = {};
+      this._requiredError = "Required";
     } else {
       this.saveFunc = options.save ? options.save : defaultSaveFunc;
       this.isDisabledFunc = options.isDisabled
@@ -196,6 +207,8 @@ export class FormState<
       this.blurFunc = options.blur ? options.blur : null;
       this.updateFunc = options.update ? options.update : null;
       this._context = options.context;
+      this._converterOptions = options.converterOptions || {};
+      this._requiredError = options.requiredError || "Required";
     }
   }
 
@@ -207,6 +220,10 @@ export class FormState<
   @computed
   get value(): Instance<M> {
     return this.node;
+  }
+
+  get stateConverterOptionsWithContext(): StateConverterOptionsWithContext {
+    return { context: this.context, ...this._converterOptions };
   }
 
   @action