+++ test/converter.test.ts
@@ -1,4 +1,6 @@
 import { configure } from "mobx";
+import { types } from "mobx-state-tree";
+import { Field, Form, converters } from "../src";
 import { CONVERSION_ERROR, ConversionValue, Converter } from "../src/converter";
 
 configure({ enforceActions: "observed" });
@@ -10,12 +12,12 @@ test("simple converter", async () => {
     render: value => value
   });
 
-  const result = await converter.convert("foo");
+  const result = await converter.convert("foo", {});
   expect(result).toBeInstanceOf(ConversionValue);
   expect((result as ConversionValue<string>).value).toEqual("foo");
 
   // the string "ConversionError" is a valid text to convert
-  const result2 = await converter.convert("ConversionError");
+  const result2 = await converter.convert("ConversionError", {});
   expect(result2).toBeInstanceOf(ConversionValue);
   expect((result2 as ConversionValue<string>).value).toEqual("ConversionError");
 });
@@ -28,11 +30,11 @@ test("converter to integer", async () => {
     render: value => value.toString()
   });
 
-  const result = await converter.convert("3");
+  const result = await converter.convert("3", {});
   expect(result).toBeInstanceOf(ConversionValue);
   expect((result as ConversionValue<number>).value).toEqual(3);
 
-  const result2 = await converter.convert("not a number");
+  const result2 = await converter.convert("not a number", {});
   expect(result2).toEqual(CONVERSION_ERROR);
 });
 
@@ -44,11 +46,11 @@ test("converter with validate", async () => {
     validate: value => value <= 10
   });
 
-  const result = await converter.convert("3");
+  const result = await converter.convert("3", {});
   expect(result).toBeInstanceOf(ConversionValue);
   expect((result as ConversionValue<number>).value).toEqual(3);
 
-  const result2 = await converter.convert("100");
+  const result2 = await converter.convert("100", {});
   expect(result2).toEqual(CONVERSION_ERROR);
 });
 
@@ -67,8 +69,33 @@ test("converter with async validate", async () => {
     render: value => value
   });
 
-  const result = converter.convert("foo");
+  const result = converter.convert("foo", {});
   done[0]();
   const v = await result;
   expect((v as ConversionValue<string>).value).toEqual("foo");
+});
+
+test("converter maybeNull with converter options", async () => {
+  const M = types.model("M", {
+    foo: types.maybeNull(types.string)
+  });
+
+  const form = new Form(M, {
+    foo: new Field(converters.maybeNull(converters.decimal()))
+  });
+
+  const o = M.create({ foo: "36365.21" });
+
+  const state = form.state(o, {
+    converterOptions: {
+      decimalSeparator: ",",
+      thousandSeparator: ".",
+      renderThousands: true
+    }
+  });
+  const field = state.field("foo");
+  await field.setRaw("36.365,20");
+  expect(field.error).toBeUndefined();
+  expect(field.raw).toEqual("36.365,20");
+  expect(field.value).toEqual("36365.20");
 });